Add common Maven build workflow (#1)
Testing with the latest version of version requires regular update to all Plexus repositories'
GitHub Actions workflows. This is a repetitive, time consuming task.

Given that the build workflow for the various Plexus project is almost the same,
we can re-use common workflow. Then using the latest version of Java is just a matter
of changing the build matrix in single file.